+++ packages/nextra/loader.js
@@ -13,7 +13,7 @@ module.exports = function (code) {
   const callback = this.async()
 
   // Note that `import()` caches, so this should be fast enough.
-  import('./dist/loader.mjs').then(mod =>
+  import('./dist/server/loader.mjs').then(mod =>
     mod.default.call(this, code, callback)
   )
 }
